pds-app-auth-spike/requirements.txt
PyJWT[crypto]
setuptools_rust
requests
